Parser V3 is complete
Completed parser V3, ready to move it to master -- this version comes with considerable speed boosting, especially on large files. Note that it now uses Numpy so that will need to be included in the python environment. Ready to be moved to master, moving the current files in master to a 'legacy' version.